requirements/base.txt
django
django-allauth
psycopg2
djangorestframework
django-filter